# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Uni-Green API",
    "description": "Contract-first API for the Uni-Green B2B sales platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/public/categories": {
      "get": {
        "tags": [
          "public catalogue"
        ],
        "summary": "Categories",
        "operationId": "categories_api_v1_public_categories_get",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Locale"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicCategoryResponse"
                  },
                  "title": "Response Categories Api V1 Public Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/products": {
      "get": {
        "tags": [
          "public catalogue"
        ],
        "summary": "Products",
        "operationId": "products_api_v1_public_products_get",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Locale"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 160
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Featured"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PublicProductSort",
              "default": "featured"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProductPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/products/{slug}": {
      "get": {
        "tags": [
          "public catalogue"
        ],
        "summary": "Product",
        "operationId": "product_api_v1_public_products__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Locale"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProductDetail"
                }
              }
            }
          },
          "404": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "field_errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Field Errors"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Locale": {
        "type": "string",
        "enum": [
          "vi",
          "en"
        ],
        "title": "Locale"
      },
...
      "PaginationMetadata": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PaginationMetadata"
      },
...
      "PublicCategoryResponse": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "meta_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Title"
          },
          "meta_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Description"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "description",
          "meta_title",
          "meta_description"
        ],
        "title": "PublicCategoryResponse"
      },
      "PublicMediaResponse": {
        "properties": {
          "alt_text": {
            "type": "string",
            "title": "Alt Text"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/PublicMediaVariantResponse"
            },
            "type": "array",
            "title": "Variants"
          }
        },
        "type": "object",
        "required": [
          "alt_text",
          "is_primary",
          "variants"
        ],
        "title": "PublicMediaResponse"
      },
      "PublicMediaVariantResponse": {
        "properties": {
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "width",
          "height",
          "url"
        ],
        "title": "PublicMediaVariantResponse"
      },
      "PublicProductDetail": {
        "properties": {
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "oem_available": {
            "type": "boolean",
            "title": "Oem Available"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/PublicCategoryResponse"
            },
            "type": "array",
            "title": "Categories"
          },
          "primary_media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicMediaResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "meta_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Title"
          },
          "meta_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Description"
          },
          "specifications": {
            "items": {
              "$ref": "#/components/schemas/PublicSpecificationResponse"
            },
            "type": "array",
            "title": "Specifications"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/PublicMediaResponse"
            },
            "type": "array",
            "title": "Media"
          }
        },
        "type": "object",
        "required": [
          "sku",
          "slug",
          "name",
          "summary",
          "oem_available",
          "featured",
          "categories",
          "primary_media",
          "description",
          "meta_title",
          "meta_description",
          "specifications",
          "media"
        ],
        "title": "PublicProductDetail"
      },
      "PublicProductPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicProductSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetadata"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "PublicProductPage"
      },
      "PublicProductSort": {
        "type": "string",
        "enum": [
          "featured",
          "name",
          "newest"
        ],
        "title": "PublicProductSort"
      },
      "PublicProductSummary": {
        "properties": {
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "oem_available": {
            "type": "boolean",
            "title": "Oem Available"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/PublicCategoryResponse"
            },
            "type": "array",
            "title": "Categories"
          },
          "primary_media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicMediaResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "sku",
          "slug",
          "name",
          "summary",
          "oem_available",
          "featured",
          "categories",
          "primary_media"
        ],
        "title": "PublicProductSummary"
      },
      "PublicSpecificationResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "is_highlighted": {
            "type": "boolean",
            "title": "Is Highlighted"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "value",
          "unit",
          "is_highlighted"
        ],
        "title": "PublicSpecificationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}